requirements.txt
langchain 
pypdf
pinecone-client
openai
tiktoken
pyPDF2
streamlit
python-dotenv
chromadb 
sentence-transformers 
spacy
transformers
transformers[sentencepiece]